rename generateUuid file by helpers
